camera bone for viewmodels
!!!IMPORTANT!!!
To use this you need to have two bones in your viewmodel, camera_root and camera. camera_root IS NOT animated and should not move at all. The only thing that should move is the camera bone or this will not work. To preview the animation in your animation software create a camera and make that a child of the camera bone, not the camera_root bone.